Refactoring/use-pageable (#52)

+++ chip-service/src/main/java/com/rljj/chipservice/domain/post/service/PostServiceImpl.java
@@ -2,20 +2,20 @@ package com.rljj.chipservice.domain.post.service;
 
 import com.rljj.chipservice.domain.post.dto.PostRequest;
 import com.rljj.chipservice.domain.post.repository.PostRepository;
+import com.rljj.switchswitchcommon.exception.NotFoundException;
 import com.rljj.switchswitchentity.chip.chippost.ChipPost;
 import com.rljj.switchswitchentity.chip.chipinfo.ChipInfo;
 import com.rljj.switchswitchentity.member.Member;
 import com.rljj.switchswitchentity.chip.chippost.ChipPostStatus;
 //import com.rljj.chipservice.domain.post.exception.PostNotFoundException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import java.util.List;
+import org.springframework.data.domain.Sort;
 
-// null 체크, setter 쓰지 않기 등  
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Service
@@ -23,9 +23,15 @@ public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
 
     @Override
-    public List<ChipPost> getPosts(int page, int size) {
-        Pageable pageable = PageRequest.of(0, 10, Sort.by("createdDate").descending());
-        return postRepository.findAll(pageable).getContent();
+    public Page<ChipPost> getPosts(int page, int size) {
+        Pageable pageable = PageRequest.of(page - 1, size, Sort.by("createdDate").descending());
+        return postRepository.findAll(pageable);
+    }
+
+    @Override
+    public ChipPost getPost(Long id) {
+        return postRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Post not found with id: " + id));
     }
 
     @Override
@@ -62,18 +68,15 @@ public class PostServiceImpl implements PostService {
                 request.getStatus()
         );
 
-        return postRepository.save(existingPost);
-    }
-
-    @Override
-    public ChipPost getPostById(Long id) {
-        return postRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Post not found with id: " + id)); //밑줄 왜 생겨?
+        return existingPost;
     }
 
     @Override
     @Transactional
     public void deletePost(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new NotFoundException(String.valueOf(id));
+        }
         postRepository.deleteById(id);
     }
 }
